selectMemoFromListを作成し実装しました
