+++ 04.class/memo.js
@@ -2,10 +2,13 @@ import readline from "readline";
 import enquirer from "enquirer";
 
 import MemoOperation, { db } from "./memoOperation.js";
+import OperateInterface from "./operateInterface.js";
 
 const { Select } = enquirer;
 
 const memoOperation = new MemoOperation();
+const operateInterface = new OperateInterface();
+
 const firstArg = process.argv[2];
 
 if (firstArg === "-l") {
@@ -15,27 +18,15 @@ if (firstArg === "-l") {
     });
   });
 } else if (firstArg === "-r") {
-  db.all("SELECT * FROM memo", (err, rows) => {
-    const memo = rows.map((row) => {
-      const firstLine = row.content.split("\n")[0];
-      return { name: firstLine, value: row.content };
-    });
-    const prompt = new Select({
-      type: "select",
-      name: "name",
-      message: "choose your memo",
-      choices: memo,
-      result() {
-        return this.focused.value;
-      },
-    });
-    prompt
-      .run()
-      .then((result) => console.log(result))
-      .catch(console.error);
+  operateInterface.selectMemoFromList((result) => {
+    console.log(result);
   });
 } else if (firstArg === "-d") {
   db.all("SELECT * FROM memo", (err, rows) => {
+    if (err) {
+      console.error(err);
+      return;
+    }
     const memo = rows.map((row) => {
       const firstLine = row.content.split("\n")[0];
       return { name: firstLine, value: row.content };
